pyproject.toml
[project]
name = "energy-engineering-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.2.1",
    "apscheduler>=3.11.2",
    "flask==3.0.2",
    "flask-cors==4.0.0",
    "pandas>=3.0.2",
    "pymongo[srv]>=4.17.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
